pyproject.toml
[project]
name = "tuna-r"
version = "0.1.0"
description = "The improved codebase for TUnA, a sequence-based protein-protein interaction classifier."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "biopython>=1.86",
    "einops>=0.8.1",
    "fair-esm>=2.0.0",
    "huggingface-hub>=1.0.1",
    "hydra-core>=1.3.2",
    "isort>=6.0.1",
    "lightning>=2.5.2",
    "pre-commit>=4.2.0",
    "pytest>=8.4.1",
    "pytorch-optimizer>=3.7.0",
    "ruff>=0.12.2",
    "torch>=2.7.1",
    "torchmetrics>=1.7.3",
    "uncertainty-calibration>=0.1.4",
    "uncertaintyawaredeeplearn",
    "wandb>=0.20.1",
]

[project.scripts]
tuna = "tuna.inference.cli:main"

[tool.uv.sources]
uncertaintyawaredeeplearn = { git = "https://github.com/jlparkI/uncertaintyAwareDeepLearn", rev = "0.0.5" }
[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
lint.ignore = ["E501"]
lint.extend-select = ["E", "F", "I"]

[dependency-groups]
dev = [
    "ipykernel>=7.2.0",
]
